started diving app in components, improved responsiveness
